FridgeViewController : added prepare for segue to know if should display favorite
RecipeDetailsViewController : - Outlet for favorite and button to favorite a recipe.

RecipeListViewController : Changed the way to display the table view to adapt it with a favorite Recipe List.

CoreDataManager : Added to manage data.

RecipeService started code for Favorite

Moved some code of CoreData from App delegate to core data file

Started test for CoreData